integrations: replacing noy-black functionality inside noy itself
This makes it so a pth file is installed in your environment that will
modify black on startup if NOSE_OF_YETI_BLACK_COMPAT=true is in the
environment

+++ tests/test_formatting.py
@@ -26,9 +26,10 @@ class TestBlackFormatting:
 
         from black.cache import CACHE_DIR
 
-        for fle in CACHE_DIR.iterdir():
-            if "noy" in fle.name:
-                fle.unlink()
+        if CACHE_DIR.exists():
+            for fle in CACHE_DIR.iterdir():
+                if "noy" in fle.name:
+                    fle.unlink()
 
         directory = tmp_path_factory.mktemp("files")
